# server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SosyApp Api",
    "description": "Sosyal medya uygulaması dokümantasyonu",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/api/posts/feed": {
      "get": {
        "description": "Aktif kullanıcının takip ettiği kişilerin gönderilerini (feed) listeler (sayfalama ile).",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          }
          
        ],
        "responses": {
          "200": {
            "description": "Gönderi akışı.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PostResponse"
              }
            }
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
    "/api/posts": {
      "post": {
        "description": "Yeni bir gönderi oluşturur.",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          },
           {
             "name": "body",
             "in": "body",
             "description": "Yeni gönderi içeriği",
             "required": true,
             "schema": {
               "$ref": "#/definitions/PostCreateInput"
             }
           }
        ],
        "responses": {
          "201": {
            "description": "Created - Gönderi oluşturuldu",
            "schema": {
              "$ref": "#/definitions/PostResponse"
            }
          },
          "400": {
            "description": "Bad Request - Geçersiz istek (örn. içerik eksik)",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
    "/api/posts/{postId}": {
      "get": {
        "description": "Belirtilen ID'ye sahip gönderiyi getirir.",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": "Getirilecek gönderinin ID'si."
          }
        ],
        "responses": {
          "200": {
            "description": "Gönderi detayları.",
            "schema": {
              "$ref": "#/definitions/PostResponse"
            }
          },
          "404": {
            "description": "Not Found - Gönderi bulunamadı",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      },
      "delete": {
        "description": "Belirtilen ID'ye sahip gönderiyi siler (Sadece gönderi sahibi silebilir).",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": "Silinecek gönderinin ID'si."
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          }
        ],
        "responses": {
          "200": {
            "description": "Gönderi başarıyla silindi."
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim veya gönderi sahibi değil",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
           "404": {
             "description": "Not Found - Gönderi bulunamadı",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
           },
          "500": {
            "description": "Internal Server Error",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
    "/api/posts/user/{userId}": {
      "get": {
        "description": "Belirtilen kullanıcı UID'sine ait tüm gönderileri getirir.",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
             "format": "uuid",
            "description": "Gönderileri getirilecek kullanıcının UID'si."
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi (isteğe bağlı, özel profiller için)."
          }
           
        ],
        "responses": {
          "200": {
            "description": "Kullanıcının gönderileri.",
             "schema": {
               "type": "array",
               "items": {
                 "$ref": "#/definitions/PostResponse"
               }
             }
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim (eğer profil gizliyse ve takip edilmiyorsa)",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Not Found - Kullanıcı veya gönderiler bulunamadı",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Internal Server Error",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/api/likes/post/{postId}": {
      "get": {
        "description": "Belirtilen gönderiyi beğenen kullanıcıların listesini getirir.",
        "tags": [
          "Likes"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": "Beğenenleri listelenecek gönderinin ID'si."
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi (isteğe bağlı)."
          }
       
        ],
        "responses": {
          "200": {
            "description": "Beğenen kullanıcıların listesi.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserResponse" 
              
              }
            }
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim (Gerekliyse)",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Not Found - Gönderi bulunamadı",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "500": {
            "description": "Internal Server Error",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
...
    "/api/follows/followers": {
      "get": {
        "description": "Mevcut kullanıcıyı takip eden kullanıcıların listesini getirir.",
        "tags": [
          "Follows"
        ],
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          }
           
        ],
        "responses": {
          "200": {
            "description": "Takipçi listesi.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserResponse" 
                
              }
            }
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
    "/api/follows/following": {
      "get": {
        "description": "Mevcut kullanıcının takip ettiği kullanıcıların listesini getirir.",
        "tags": [
          "Follows"
        ],
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          }
         
        ],
        "responses": {
          "200": {
            "description": "Takip edilen listesi.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserResponse"
                
              }
            }
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          }
        }
      }
    },
    "/api/follows/{userId}": {
      "post": {
        "description": "Mevcut kullanıcının belirtilen kullanıcıyı takip etmesini sağlar.",
        "tags": [
          "Follows"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": "Takip edilecek kullanıcının UID'si."
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          }
        ],
        "responses": {
          "200": {
            "description": "Kullanıcı başarıyla takip edildi.",
            "schema": {
              "$ref": "#/definitions/FollowResponse" 
            }
          },
          "400": {
            "description": "Bad Request - Kendini takip edemezsin",
            "schema": { "$ref": "#/definitions/ErrorResponse" }
           },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
           "404": {
             "description": "Not Found - Takip edilecek kullanıcı bulunamadı",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
           },
           "409": {
             "description": "Conflict - Kullanıcı zaten takip ediliyor",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
           }
        }
      },
      "delete": {
        "description": "Mevcut kullanıcının belirtilen kullanıcıyı takipten çıkmasını sağlar.",
        "tags": [
          "Follows"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
             "format": "uuid",
            "description": "Takipten çıkılacak kullanıcının UID'si."
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          }
        ],
        "responses": {
          "200": {
            "description": "Kullanıcı başarıyla takipten çıkarıldı."
          },
           "400": {
             "description": "Bad Request - Kendini takipten çıkamazsın",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
           },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
             "description": "Not Found - Kullanıcı veya takip ilişkisi bulunamadı",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
           }
        }
      }
    },
    "/api/follows/follower/{followerId}": {
      "delete": {
        "description" : "Mevcut kulanicinin belirtilen takipcini siler.",
        "tags": [
          "Follows"
        ],
        "parameters" : [
          {
            "name": "followerId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid",
            "description": "Takipcini silinecek kullanıcının UID'si."
          },
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          }
        ],
        "responses" : {
          "200": {
            "description": "Kullanıcı başarıyla takipcilerden silindi."
          },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
             "description": "Not Found - Kullanıcı veya takipcilik ilişkisi bulunamadı",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
           }
        }
      }
    },
...
    "/api/search/posts": {
      "get": {
        "description": "Belirtilen sorgu parametresine göre gönderileri arar (içerik).",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "cookie",
            "in": "header",
            "type": "string",
            "description": "Kimlik doğrulama için kullanıcı oturum çerezi."
          },
          {
            "name": "query",
            "in": "query",
            "type": "string",
            "required": true,
            "description": "Arama yapılacak metin (gönderi içeriği)."
          }
        ],
        "responses": {
          "200": {
            "description": "Eşleşen gönderilerin listesi.",
            "schema": {
               "type": "array",
               "items": {
                 "$ref": "#/definitions/PostResponse"
               }
             }
          },
           "400": {
              "description": "Bad Request - Sorgu parametresi eksik",
              "schema": { "$ref": "#/definitions/ErrorResponse" }
            },
          "403": {
            "description": "Forbidden - Yetkisiz erişim",
             "schema": { "$ref": "#/definitions/ErrorResponse" }
          },
          "404": {
            "description": "Not Found - Eşleşen gönderi bulunamadı (Boş liste dönebilir)",
            "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PostResponse"
                }
              }
          }
        }
      }
    }
  },
  "definitions": {
...
    "UserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "readOnly": true,
          "description": "Veritabanı ID"
        },
        "uid": {
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "description": "Benzersiz kullanıcı ID"
        },
        "role": {
          "type": "string",
          "readOnly": true,
          "default": "user",
          "example": "user"
        },
        "username": {
          "$ref": "#/definitions/UserBase/properties/username"
        },
        "email": {
          "$ref": "#/definitions/UserBase/properties/email"
        },
        "firstName": {
          "$ref": "#/definitions/UserBase/properties/firstName"
        },
        "lastName": {
          "$ref": "#/definitions/UserBase/properties/lastName"
        },
        "profilePicture": {
          "$ref": "#/definitions/UserBase/properties/profilePicture"
        },
        "bio": {
          "$ref": "#/definitions/UserBase/properties/bio"
        },
        "location": {
          "$ref": "#/definitions/UserBase/properties/location"
        },
        "verified": {
          "type": "boolean",
          "readOnly": true,
          "default": false,
          "example": false
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        }
      }
    },
...
    "PostResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "readOnly": true
        },
        "userId": {
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "description": "Gönderiyi oluşturan kullanıcının UID'si"
        },
        "content": {
          "$ref": "#/definitions/PostBase/properties/content"
        },
        "media": {
          "$ref": "#/definitions/PostBase/properties/media"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "User": {
          "$ref": "#/definitions/UserResponse"
        },
        "likesCount": {
          "type": "integer",
          "example": 15,
          "readOnly": true
        },
        "commentsCount": {
          "type": "integer",
          "example": 3,
          "readOnly": true
        }
      }
    },
    "PostCreateInput": {
      "type": "object",
      "required": [
        "content"
      ],
      "properties": {
        "content": {
          "$ref": "#/definitions/PostBase/properties/content"
        },
        "media": {
          "$ref": "#/definitions/PostBase/properties/media"
        }
      }
    },
...
    "FollowResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "readOnly": true
        },
        "followerId": {
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "description": "Takip eden kullanıcının UID'si"
        },
        "followingId": {
          "type": "string",
          "format": "uuid",
          "readOnly": true,
          "description": "Takip edilen kullanıcının UID'si"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "FollowerUser": {
          "$ref": "#/definitions/UserResponse"
        },
        "FollowedUser": {
          "$ref": "#/definitions/UserResponse"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Hata mesajı",
          "example": "Kullanıcı bulunamadı."
        }
      },
      "required": [
        "message"
      ]
    }
  }
}